👌 move RUM entry points
