+++ eslint-local-rules/disallowSideEffects.js
@@ -28,8 +28,7 @@ const packagesRoot = path.resolve(__dirname, '..', 'packages')
 const pathsWithSideEffect = new Set([
   `${packagesRoot}/logs/src/boot/logs.entry.ts`,
   `${packagesRoot}/logs/src/index.ts`,
-  `${packagesRoot}/rum/src/boot/rum.entry.ts`,
-  `${packagesRoot}/rum/src/index.ts`,
+  `${packagesRoot}/rum/src/entries/main.ts`,
   `${packagesRoot}/rum-slim/src/boot/rumSlim.entry.ts`,
   `${packagesRoot}/rum-slim/src/index.ts`,
 ])
