+++ packages/rum-core/src/domain/rumEventsCollection/action/trackActions.spec.ts
@@ -1,8 +1,8 @@
 import { Context, DOM_EVENT, ClocksState, Observable } from '@datadog/browser-core'
 import { Clock, createNewEvent } from '../../../../../core/test/specHelper'
-import { RumEvent } from '../../../../../rum/src'
 import { setup, TestSetupBuilder } from '../../../../test/specHelper'
 import { RumEventType, ActionType } from '../../../rawRumEvent.types'
+import { RumEvent } from '../../../rumEvent.types'
 import { LifeCycleEventType } from '../../lifeCycle'
 import { PAGE_ACTIVITY_VALIDATION_DELAY } from '../../waitIdlePage'
 import { AutoAction, AUTO_ACTION_MAX_DURATION, trackActions } from './trackActions'
