+++ packages/rum/internal/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "main": "../cjs/internal.js",
-  "module": "../esm/internal.js",
-  "types": "../cjs/internal.d.ts"
+  "main": "../cjs/entries/internal.js",
+  "module": "../esm/entries/internal.js",
+  "types": "../cjs/entries/internal.d.ts"
 }
