+++ packages/rum/src/entries/internal.ts
@@ -19,7 +19,7 @@ export {
   AddedNodeMutation,
   MousePosition,
   RemovedNodeMutation,
-} from './domain/record/types'
+} from '../domain/record/types'
 
 export {
   PRIVACY_ATTR_NAME,
@@ -28,8 +28,8 @@ export {
   PRIVACY_CLASS_HIDDEN,
   PRIVACY_CLASS_INPUT_IGNORED,
   NodePrivacyLevel,
-} from './constants'
+} from '../constants'
 
-export * from './types'
+export * from '../types'
 
-export { serializeNodeWithId } from './domain/record/serialize'
+export { serializeNodeWithId } from '../domain/record/serialize'
